Update activity_home_dashboard_fragment.xml
Got problems connecting to the bar chart package. Is it on my comp or is it the same for you also? I followed the steps exactly, and also a few other tutorials I've tried. We have that extra 'plug-in' in the build Gradle that's different from the tutorial. Not sure if that's what causing the problem though I don't think it should be.